🧹 [code health improvement] Use shared scene-parser in scenes.ts
